+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Unreleased
 
+- Anthropic: Support for images with mime type image/bmp.
 - Sandbox Service: Specify instance externally so a single script can service multiple instances.
+- Agent Bridge: Embed sandbox service client in sandbox bridge proxy (for ease of bundling).
 - Sandboxes: Add Sandbox Injection API for Dynamic Environment Setup.
   
 ## 0.3.131 (08 September 2025)